requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
flake8==7.3.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg==3.2.9
psycopg-binary==3.2.9
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
uritemplate==4.2.0
